backend-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
whisper==1.1.10
PyMuPDF==1.23.8
sentence-transformers==2.2.2
chromadb==0.4.18
minio==7.2.0
httpx==0.25.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
prisma==0.11.0

